Split /me/api from internal calls to websockets
/me/api has so far been a ugly hybrid of sometimes cookie authenticated
and sometimes OAuth, using unsafe (if correct) Origin checks, and LGTM
has rightly complained.

Instead, create two endpoints:
- /me/ws/conversation to call from the browser in almond-cloud pages
  (cookie authenticated, with Origin/CSRF protection)
- /me/api/conversation to call from everywhere else, with OAuth
  authentication and no Origin or CSRF protection (because OAuth)
