+++ views/my_conversation.pug
@@ -11,7 +11,7 @@ block page_name
   = _("Web Almond")
 
 block content
-  div#conversation.single(data-target='/me/api/conversation')
+  div#conversation.single(data-target='/me/ws/conversation')
     div#chat
     div#input-block
       span(role='button').hidden#cancel
