+++ views/my_stuff.pug
@@ -101,7 +101,7 @@ block content
 
   br
   br
-  div#conversation.single(data-target='/me/api/conversation')
+  div#conversation.single(data-target='/me/ws/conversation')
     div#chat
     div#input-block
       span(role='button').hidden#cancel
