+++ views/stanford/about_get_almond.pug
@@ -27,7 +27,7 @@ block content
 
             a(href='#section-try-now').btn.btn-primary.col-xs-8.col-xs-offset-2.col-md-4.col-md-offset-4#try-almond-btn= _("Try Almond")
 
-            div#conversation(data-target='/me/api/anonymous',aria-expanded='false').collapse
+            div#conversation(data-target='/me/ws/anonymous',aria-expanded='false').collapse
               div#chat
               div#input-block
                 span(role='button').hidden#cancel
